+++ src/command/scrub/setup.rs
@@ -9,10 +9,10 @@ use crate::{
         replicate::ReplicationChecker,
         scrub::{
             check::{
-                BlobChecker, BlobIndexChecker, DigestLinkChecker, LayoutChecker,
-                LinkReferencesChecker, ManifestChecker, MediaTypeChecker, MultipartChecker,
-                NamespaceChecker, OrphanGrantChecker, OrphanJobChecker, OrphanNamespaceChecker,
-                OrphanQueue, ReferrerChecker, StoreChecker, TagChecker, UploadChecker,
+                BlobChecker, BlobIndexChecker, DigestLinkChecker, LinkReferencesChecker,
+                ManifestChecker, MultipartChecker, NamespaceChecker, OrphanGrantChecker,
+                OrphanJobChecker, OrphanNamespaceChecker, OrphanQueue, ReferrerChecker,
+                StoreChecker, TagChecker, UploadChecker,
             },
             command::Options,
             error::Error,
@@ -82,13 +82,6 @@ pub fn namespace_checkers(
         )));
     }
 
-    if options.media_types {
-        checkers.push(Box::new(MediaTypeChecker::new(
-            blob_store.clone(),
-            metadata_store.clone(),
-        )));
-    }
-
     if options.referrers {
         checkers.push(Box::new(ReferrerChecker::new(metadata_store.clone())));
     }
@@ -117,10 +110,6 @@ pub fn tag_checkers(
             metadata_store.clone(),
         )) as Box<dyn TagChecker>]
     })
-}
-
-pub fn layout_checker(blob_store: &Arc<BlobStore>) -> LayoutChecker {
-    LayoutChecker::new(blob_store.clone())
 }
 
 pub fn blob_checker(